renault-detector/requirements.txt
# Dependências principais para o projeto Renault Detector
ultralytics>=8.0.0
opencv-python>=4.8.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0

# Dependências para download de imagens
duckduckgo-search>=3.9.0
requests>=2.31.0

# Dependências para visualização e análise
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Dependências para monitoramento
wandb>=0.15.0

# Dependências para processamento de dados
pyyaml>=6.0
pathlib2>=2.3.0

# Dependências opcionais para desenvolvimento
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

